Adds number_possible_a function
Change the way to determine the number of options available. Adds a
continous function that maps the state in the number of available
options, later converting to integer. Now we first compute the number of
possible options, and later return the options available, changing the
function _possible_a. This was made to allow more flexibility and make
it easier to alter the state -> options function.