Refactor code to improve variable names
The naming was confusing and incompatible with previous version of the
work in another documents. We now opt for using P to describe the main
state (instead of x), and 'I' or influx instead of 'a'. We also
reorganize the order of some functions, removed unused functions and
add the differential equation on the notebook.