Refactor code to use two steps
Before I there was a step variable that was used to evolve the EDO and
compute pathway diversity. This is confusing and these steps in
principle can be different. It refactors the code to have two variables
the time step, and the decision step. The first defines the interval of
data collection of the state and causal entropy. The later defines the
steps used in the causal entropy. It also change the variable name
maximun time in run_entropy to number of decisions.