Refactor notebook plots
The plots were disorganized and with uncessary code. It refactors the
plots to do not use loops, using the original data directly, which
largely reduced the code. It also adds a plot with multiple subplots for
the early-warning signal analysis. It also adds a way to filter the data
to remove series on the plot step.