Adds new file metrics
The computation of variance and autocorrelation was ocurring inside the
notebook. To start removing logic from the notebook, this commit
extracted this code to a new file metrics.jl.
It also fix a bug in the choice of time in the variance and
autocorrelation computation. The code was not overlaping the sliding
windows. This commit also fix this behaviour. Now the size of the
variance time series is similar to the state time series, removing the
initial interval.