Fix _entropy using wrong influx value
The _entropy code was always using the same value for the influx to
evolve the time step. It should use the values according to the
_possible_influx choosen.
This was happening because of a wrong variable name, it was always using
the function input I instead of the value from the _possible_influx map.
This commit change the variable name to use the correct value.