Adds a procedural version of the entropy computation
This is part of an analysis done to try improvind the performance of the
entropy method. In general procedural version consumes less memory and
run faster than recursive methods. However, in this case the procedural
is taking longer to run. So this code was abandoned for now and kept
here just in case further analysis and improvements want to use it.