Normalise entropy in the scalling simulation
After some investigations to find a normalisation based on the maximum
number of options, it was found that the maximum causal entropy depends
on the number of decisions times the log of the maximum number of
options. This commit implements the normalisation in the scalling
results. Possibly it will be added for the other results as well.