Modify #_entropy tests to be more explicit
The #_entropy tests were using numerical values to check the results of
the #_entropy function, but this is not a good practice. It modify the
tests to use a formula using logs, based on a manual computation of the
results. Only when using the function with higher number of decisions it
is still using a numeric value.