Make EWS analysis use increasing influx
In the early warning signal analysis it was using a fixed influx,
varying this value to get the necessary behaviours. This is not adequate
to make EWS analysis because we need the increase behaviour to happen
slowly and progressively. This commit changes to start in a common
influx and then having different tax of influx increase.