Implement array of results in EWS analysis
It was hard to operate all the results, so to make a more coherent
structure and prepare for the use of NamedDims it creates an array of
results output by the #run_scenarios function. This lead to a rewrite of
other parts of the code that use this output, such as the
function #early_warning_signals of the cookbook and the metrics.

Another change was implemented in the #run_scenarios function. Now it
does receive the number of scenarious, as it was always being runned
with one scenario. Insted, it receives the varying parameter I_taxes.