Change variable parameter in EWS analysis
Before in the EWS analysis the rate of influx per time was the variable
parameter. However, based on discussions with co-authors we decided a
more interesting parameter is the time frame. As this parameter is only
used of the pathway diversity (and not the state). This lead to some
bigger changes in the code. Especifically, now p only depends on time
and not on a secondary parameter.