Adds writing result in CSV
To decouple the model from transformation and analysis it saves a csv
with the model results. This will help chaging transformations of the
data without having to run the model again and also helps with
reproductibility, especially for the EWS part. Later the code will read
the csv instead of receiving the result.