Change computation of kendall-tau
Before it was selecting the start point to compute the kendall-tau.
Varying over the time series to select the maximum value.
Now it is computing the end point to compute the kendall-tau and we
change the analysis to compute the time where it passes a defined
threshold.

The analysis still has some problems that will be changed soon.