Removes kendall-tau value from EWS plot
The EWS plot changed and now we plot the point in time where you can
stop the time series and still get a relevant kendall-tau. So, we adopt
a kendall-tau threshold. As a result, all kendall-tau plot have the same
valuye and it is not longer necessary. So, it removes the kendall-tau
value from the label of each line in the EWS plot.