Rename number_options to max_number_options in base.jl
The name used before was confusing. Number of options measure how many
options you have given a state. The maximum number of options is
a parameter of the model and constant over time. The functions #_entropy
and #_possible_influx was using the term number_options, but the meaning
was really max_number_options. This commit change the variable name to
keep consistency.