Adds more time_horizons in DT analysis
To have fewer lines we removed option diversity from the distance
to threshold analysis. Now we reconsider this because we want to have
this baseline to compare and argue about the necessity of pathway
diversity. We also change the initialisation of DT parameters to define
time_horizon in terms of the number of decisions being made and the
decision scale.