Adds new possibility of entropy normalisation
Based on a theoretical result we can normalise pathway diversity based
on the number of decisions and maximum number of options. This commit
implements this possiblity.
This commit also uses this new normalisation in the DT plot. However, it
is not used in the EWS plot, because the results are not as expected.
Futurely need to look why the normalisation is not giving the expected
result.