Change _entropy to use keyword arguments
The way _entropy is designed makes more sense with keyword arguments
because argumets are optional with default values. This will also
prepare to add one more keyword arguments. The commit makes the code
easier to read, because otherwise optional parameters have to be passed
in order.
It also change the name of one of the optional arguments to have a
shorter argument list.