Fix possible_influx optional args
On commit 53b6ede6 it implemented optional arguments maximum and minimum
influx on _entropy and it pass the values to _possible_influx. However,
one function uses the _possible_influx directly, so it necessary to
include maximum and minimum influx default value on this function too.
It also fix a spelling problem, it was using maximun instead of maximum.