Change default parameters in distance to threshold and sensitivity
The final value for P0 in the distance to threshold parameter is too
high a deforms the final result. Following what was adopted in the
sensitivity plot, it changes the parameter P0, to make the rand final
value smaller.

The initial influx is not important generally. However, it is very
relevant on the minimun change scenario sensitivity analysis. To produce
a result that makes more sense for this scenario, we changed the default
value of the influx.