Refactor sensitivity analysis parameter calling
Before it defined a dictionary with all possible parameters for all
scenarios. But with the number of scenarios growing the number of
parameters can also grow. To simplify it, now we include name and
decision step for all scenario and all other optional arguments is only
included if the scenario changed its default value.