Update parameters for influx limits and update entropy tests
The parameters for influx limits were not making sense, it was using
values that are too high (0.36) and this was deforming the results. It
changes the default value to 0.3 and change scenario variations
accordingly. It refactor the tests and update it to account for the new
parameters. It also adds missing tests for keyworkd arguments of the
_entropy function that were not being tested before.