Implement new pathway diversity computation
The current method to implement pathway diversity is too slow and has
problems to increase resolution. It implement the new method to compute
pathway diversity deterministicly. Currently it is not being used in
other methods and tests are being made to guarantee the results are
consistent. The function accepts the same parameters.