Refactor probability computation
We will add new methods to compute the probability transition.
Therefore, it extractred the functions that chooses which method will be
used to a new functions. This was done with both the new and old method.
We also renamed the possible_a_vec to possible_influx, to reflect the
most uptodate nomenclature used on the code.