Update find peaks method in distance to threshold analysis
Before the code was taking the original pathway diversity, computing the
derivative and then detrending with the loess algorithm and finding
peaks. The result was not good and it was hard to follow that the peak
was really what the algorithm had compute. It changes the code to
detrend with the loess algorithm the original pathway diversity result,
then compute the derivative and finding peaks. The result improved and
the visualisation is clear on where are the peaks. It also adds a filter
to only consider peaks with a certain prominance (see Peaks docs for
more details)