Change default parameters for state distribution plot
Now the code is showing the range of the threshold. It would make more
sense for the plot to choose states outside of the threshold region. So
this commit change the default parameters for the initial states.
Also now the plot is showing the state directly and not the distance to
threshold in the title, so we choose states with less decimals.