Change normalisation for EWS analysis
Before the code was normalising pathway diversity by the maximum value.
This produces a better visualisation, but is not compatible with the
normalisation by maximum number of pathways done in the rest of the
analysis. To improve coherence, this commit changes the EWS pathway
diversity normalisation to follow the standard approach.