Compute kendall_tau at the threshold
Previously we did an analysis where we defined a threshold and find out
the point in time where the kendall-tau crossed this threshold for the
first time. This is not the standard approach to use kendall-tau and we
changed this analysis to compute the kendall-tau with time-series up to
the threshold point minus an offset, reporting this value. The offset is
determined by when the residuals start deviating from the mean value of
zero.