Change default parameters of sensitivity analysis
To make the explanation of the scenarios easier, this commit makes
changes in the default parameters used in the sensitivity analysis. We
changed the parameters of the minimal change scenario and favour lower
influx. The favour lower influx was using incorrect parameters, it
should keep the range of influx, but was actually increasing. We change
it to keep maximum value and reduce the minimum value.